hw8/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
apache-beam==2.51.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cloudevents==1.10.0
cloudpickle==2.2.1
contourpy==1.2.0
crcmod==1.7
cycler==0.12.1
deprecation==2.1.0
dill==0.3.1.1
dnspython==2.4.2
docopt==0.6.2
fastapi==0.103.2
fastavro==1.9.0
fasteners==0.19
Flask==2.3.3
fonttools==4.44.0
functions-framework==3.4.0
google-api-core==2.12.0
google-apitools==0.5.31
google-auth==2.23.2
google-auth-httplib2==0.1.1
google-cloud-aiplatform==1.36.2
google-cloud-bigquery==3.13.0
google-cloud-bigquery-storage==2.22.0
google-cloud-bigtable==2.21.0
google-cloud-core==2.3.3
google-cloud-datastore==2.18.0
google-cloud-dlp==3.13.0
google-cloud-language==2.11.1
google-cloud-pubsub==2.18.4
google-cloud-pubsublite==1.8.3
google-cloud-recommendations-ai==0.10.5
google-cloud-resource-manager==1.10.4
google-cloud-spanner==3.40.1
google-cloud-storage==2.11.0
google-cloud-videointelligence==2.11.4
google-cloud-vision==3.4.5
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpc-google-iam-v1==0.12.6
grpcio==1.59.0
grpcio-status==1.59.0
gunicorn==20.1.0
h11==0.14.0
hdfs==2.7.3
httplib2==0.22.0
httptools==0.6.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Js2Py==0.74
kiwisolver==1.4.5
llvmlite==0.41.0
MarkupSafe==2.1.3
matplotlib==3.8.1
numba==0.58.0
numpy==1.24.4
oauth2client==4.1.3
objsize==0.6.1
orjson==3.9.10
overrides==6.5.0
packaging==23.1
pandas==2.1.2
Pillow==10.1.0
proto-plus==1.22.3
protobuf==4.24.3
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.4.2
pydantic_core==2.10.1
pydot==1.4.2
pyjsparser==2.7.1
pymongo==4.6.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
rsa==4.9
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.27.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.5
uvicorn==0.23.2
uvloop==0.17.0
waitress==2.1.2
watchdog==3.0.0
watchfiles==0.20.0
websockets==11.0.3
Werkzeug==3.0.0
zstandard==0.22.0
